refactor: Activities step
